requirements.txt
django>=2.2.0,<3.2
django-crispy-forms>=1.6.0
django-reversion>=3.0.0
django-formtools>=2.0
django-import-export>=0.5.1
django-ckeditor>=5.7.0
httplib2>=0.16.0
future
six
xlwt
xlsxwriter
psycopg2-binary
gunicorn
supervisor
djangorestframework
django-filter
markdown
djangorestframework-gis==0.3
